tried to pass variable length string of uint32_t

+++ src/avr/spi.c
@@ -87,7 +87,7 @@ spi_prepare(struct spi_config config)
 
 void
 spi_transfer(struct spi_config config, uint8_t receive_data
-             , uint8_t len, uint8_t *data)
+             , uint8_t len, uint32_t *data)
 {
     if (receive_data) {
         while (len--) {
